+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixes dynamic colors and monochrome mode not supporting plain black as backgrounds.
+
 ### Security
 
 ## [0.43.0] - 2025-07-05